Move widget-specific warnings to ui.error
