+++ js/ui/widget/ui.errors.js
@@ -335,4 +335,14 @@ export default errorUtils(errors.ERROR_MESSAGES, {
     * @name ErrorsUIWidgets.W1018
     */
     W1018: 'Infinite scrolling may not work properly with multiple selection. To use these features together, set \'selection.deferred\' to true or set \'selection.selectAllMode\' to \'page\'.',
+
+    /**
+    * @name ErrorsUIWidgets.W1019
+    */
+    W1019: 'Filter query string exceeds maximum length limit of {0} characters.',
+
+    /**
+    * @name ErrorsUIWidgets.W1020
+    */
+    W1020: 'hideEvent is ignored when the shading property is true'
 });
